services/price_feeder_py/requirements.txt
paho-mqtt==2.0.0
httpx==0.25.2
python-dateutil==2.8.2